moved script editor logic into different folder/files
