+++ src/script_editor/__init__.py
@@ -1,0 +1,2 @@
+from .nuke_se import NukeScriptEditor
+from .nuke_se_controllers import CodeEditor
