+++ src/utils/__init__.py
@@ -1,4 +1,3 @@
 from .util import validate_output, get_ip, insert_time
 from .settings import AppSettings
-from .get_script_editor import NukeScriptEditor, CodeEditor
 from .paths import *
